fix: 중복값 입력 예외처리 메서드 경로 변경 (Convertor -> Lotto)
