+++ src/main/java/lotto/domain/Convertor.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import static java.util.Arrays.stream;
-import static lotto.constants.ExceptionMessage.IS_DUPLICATED;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -14,7 +13,6 @@ public class Convertor {
                 .map(Number::getNumber)
                 .sorted()
                 .collect(Collectors.toList());
-        validateDuplicated(numbers);
         return numbers;
     }
 
@@ -27,17 +25,5 @@ public class Convertor {
 
     private String[] separate(String input) {
         return input.split(DELIMITER);
-    }
-
-    private void validateDuplicated(List<Integer> numbers) {
-        if (isDuplicated(numbers)) {
-            throw new IllegalArgumentException(IS_DUPLICATED.getMessage());
-        }
-    }
-
-    private boolean isDuplicated(List<Integer> number) {
-        long result = number.stream()
-                .distinct().count();
-        return result != number.size();
     }
 }
