+++ src/main/java/lotto/domain/Lotto.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import static lotto.constants.ExceptionMessage.IS_DUPLICATED;
 import static lotto.constants.ExceptionMessage.IS_INSUFFICIENT;
 
 import java.util.List;
@@ -8,15 +9,27 @@ public class Lotto {
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
-        validate(numbers);
+        validateSize(numbers);
+        validateDuplicated(numbers);
         this.numbers = numbers;
     }
 
-    private void validate(List<Integer> numbers) {
+    private void validateSize(List<Integer> numbers) {
         if (numbers.size() != 6) {
             throw new IllegalArgumentException(IS_INSUFFICIENT.getMessage());
         }
     }
 
     // TODO: 추가 기능 구현
+    private void validateDuplicated(List<Integer> numbers) {
+        if (isDuplicated(numbers)) {
+            throw new IllegalArgumentException(IS_DUPLICATED.getMessage());
+        }
+    }
+
+    private boolean isDuplicated(List<Integer> numbers) {
+        long result = numbers.stream()
+                .distinct().count();
+        return result != numbers.size();
+    }
 }
